pythonDICOM/requirements.txt
python-gdcm==3.0.17.1
